requirements.txt
beproduct>=1.0.0
streamlit>=1.32.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
pandas>=2.2.0
requests>=2.31.0
